refactor(apex-lsp-compliant-services): extract wire-request types - W-24145209
Move the LSP dispatch wire-request shapes (PositionReq, DocOnlyReq,
DocWithContentReq, RefsReq, RenameReq, CompletionReq, SignatureHelpReq,
CodeActionReq) and completionResultForWire into src/worker/requestTypes.ts.

Pure structural types, so this is a leaf module. rename.ts now takes
RenameReq/PositionReq directly from the leaf instead of the composition root,
and the shared module imports the types for local use and re-exports them to
preserve its public surface.

W-24145209
